+++ internal/service/appmesh/virtual_router.go
@@ -66,7 +66,6 @@ func ResourceVirtualRouter() *schema.Resource {
 							Type:     schema.TypeList,
 							Required: true,
 							MinItems: 1,
-							MaxItems: 1,
 							Elem: &schema.Resource{
 								Schema: map[string]*schema.Schema{
 									"port_mapping": {